move params test to outside of integration folder
